requirements.txt
yt-dlp>=2024.1.1
feedparser>=6.0.10
openai-whisper>=20231117
ollama>=0.1.0
moviepy>=1.0.3
selenium>=4.15.0
python-dotenv>=1.0.0
schedule>=1.2.0
webdriver-manager>=4.0.1
flask>=3.0.0

